Struct hyperion_driver_acpi::hpet::TimerN

pub struct TimerN { /* private fields */ }

Implementations

impl TimerN

pub fn handler(&self) -> ApicId

pub fn sleep_until(&mut self, deadline: u64)

non blocking sleep, this triggers an interrupt after dur

deadline is in HPET clock ticks

if deadline is before the current tick, the interrupt never happens

pub fn init(&mut self)

pub fn config_and_caps(&mut self) -> TimerNConfigAndCaps

pub fn set_config_and_caps(&mut self, val: TimerNConfigAndCaps)

pub fn comparator_value(&mut self) -> CounterValue

pub fn set_comparator_value(&mut self, val: CounterValue)